+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+### Added
+
+- **Breaking:** Return early with an error when any affected file is uncommitted, unless `--force`d
+
 ## [0.5.0] - 2025-07-13
 
 ### Added